module: Provide no-op versions of init_module and finit_module
Traditionally, qcacld-3.0 is built as a DLKM, installed to either
/system or /vendor, depending on if the device supports Treble or not.
On pre-system-as-root devices like the OnePlus 5, we could build the
kernel module, add it to the ramdisk during installation, then bind
mount it during boot so the Wi-Fi HAL would handle loading it. The
advantage of this setup was the systemless aspect; uninstalling the
kernel only required reflashing the boot image, not the entire ROM.

Unfortunately, on system-as-root devices, the ramdisk is a part of
/system so this process no longer applies. Magisk does have the ability
to overlay files onto / via the /overlay folder but I have not been
able to get bind mounting to work properly. Additionally, I don't
believe that Magisk should be required to flash a custom kernel. We
could force push the modules to /system or /vendor but I believe that
requiring a full ROM flash to uninstall a kernel is poor user experience

The only alternative is building the Wi-Fi driver into the kernel image.
However, when the Wi-Fi driver is built as a DLKM during an Android
build, the Wi-Fi HAL expects that it will be loading the driver via the
init_module and finit_module syscalls.

https://android.googlesource.com/platform/frameworks/opt/net/wifi/+/android-8.1.0_r1/libwifi_hal/wifi_hal_common.cpp#61
https://android.googlesource.com/platform/frameworks/opt/net/wifi/+/android-9.0.0_r3/libwifi_hal/wifi_hal_common.cpp#64

If those syscalls fail for any reason, the Wi-Fi HAL errors out and
never reloads.

https://android.googlesource.com/platform/frameworks/opt/net/wifi/+/android-8.1.0_r1/libwifi_hal/wifi_hal_common.cpp#150
https://android.googlesource.com/platform/frameworks/opt/net/wifi/+/android-9.0.0_r3/libwifi_hal/wifi_hal_common.cpp#156

Luckily, we can get around this by providing no-op versions of these
syscalls so the Wi-Fi HAL thinks that loading the module was successful.
Because turning these syscalls into no-ops breaks loading modules
altogether, make this functionality depend on CONFIG_MODULES being
disabled.

Thanks to the previous qcacld-3.0 patches by Sultan and my removal of
the forced wait when the driver is built into the kernel image, the
driver loads within 2 seconds of kernel init and it is ready for the HAL
every time.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: engstk <eng.stk@sapo.pt>